Specify how to notify people who will be on duty
The text says that each Monday at noon UTC "a reminder is sent to
conduct@djangoproject.com with information who is the next primary and
secondary member on-duty following week", but does not specify who sends
that message, and as far as I can tell this isn't working. Instead, I
think it will work if the persons whose duty ends simply notify on Slack
the ones whose duty begins. New members of the committee will easily
pick up the habit and the trivial discussion will keep the channel
warmed up.